refactor: move all call related events into the same file
